// frontend/src/types/Cart.ts
export type CartItem = {
  _id: string;
  name: string;
  slug: string;
  image: string | undefined;
  price: number;
  quantity: number;
  countInStock: number;
  taste: string[]
};

export type ShippingAddress = {
  fullName: string;
  address: string;
  city: string;
  country: string;
  postalCode: string;
};

export type Cart = {
  cartItems: CartItem[];
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  itemsPrice: number
  shippingPrice: number
  taxPrice: number
  totalPrice: number
  favorites: any
};